feat: Move runtime to its package; split stringify-dependencies from it
