+++ packages/messageformat/package.json
@@ -73,7 +73,8 @@
   "dependencies": {
     "make-plural": "^5.0.0",
     "messageformat-formatters": "file:../formatters",
-    "messageformat-parser": "file:../parser"
+    "messageformat-parser": "file:../parser",
+    "messageformat-runtime": "file:../runtime"
   },
   "scripts": {
     "prebuild": "babel src -d lib",
